+++ src/common/routing_table.rs
@@ -1,7 +1,6 @@
 //! Simplified Kademlia routing table
 
 use std::collections::BTreeMap;
-use std::fmt::Debug;
 use std::slice::Iter;
 
 use crate::common::{Id, Node, MAX_DISTANCE};
@@ -31,6 +30,18 @@ impl RoutingTable {
     pub fn with_id(mut self, id: Id) -> Self {
         self.id = id;
         self
+    }
+
+    // === Getters ===
+
+    /// Returns the [Id] of this node, where the distance is measured from.
+    pub fn id(&self) -> Id {
+        self.id
+    }
+
+    /// Returns the map of distances and their [KBucket]
+    pub fn buckets(&self) -> &BTreeMap<u8, KBucket> {
+        &self.buckets
     }
 
     // === Public Methods ===
@@ -160,6 +171,12 @@ impl KBucket {
         }
     }
 
+    // === Getters ===
+
+    pub fn len(&self) -> usize {
+        self.nodes.len()
+    }
+
     // === Public Methods ===
 
     pub fn add(&mut self, incoming: Node) -> bool {
@@ -195,6 +212,10 @@ impl KBucket {
 
     pub fn is_empty(&self) -> bool {
         self.nodes.is_empty()
+    }
+
+    pub fn is_full(&self) -> bool {
+        self.nodes.len() >= MAX_BUCKET_SIZE_K
     }
 
     pub fn iter(&self) -> Iter<'_, Node> {